+++ PeripheralAudit/Program.cs
@@ -55,32 +55,32 @@ void ConfigureAppSettings(HostBuilderContext context, IServiceCollection collect
     string? property = config["PeripheralCosts:Dock"];
     if (property is null)
         throw new ArgumentNullException("PeripheralCosts:Dock not found in appsettings.json");
-    _costs.Dock = float.Parse(property);
+    _costs.Dock = decimal.Parse(property);
 
     property = config["PeripheralCosts:Monitor"];
     if (property is null)
         throw new ArgumentNullException("PeripheralCosts:Monitor not found in appsettings.json");
-    _costs.Monitor = float.Parse(property);
+    _costs.Monitor = decimal.Parse(property);
 
     property = config["PeripheralCosts:LargeMonitor"];
     if (property is null)
         throw new ArgumentNullException("PeripheralCosts:LargeMonitor not found in appsettings.json");
-    _costs.LargeMonitor = float.Parse(property);
+    _costs.LargeMonitor = decimal.Parse(property);
 
     property = config["PeripheralCosts:Keyboard"];
     if (property is null)
         throw new ArgumentNullException("PeripheralCosts:Keyboard not found in appsettings.json");
-    _costs.Keyboard = float.Parse(property);
+    _costs.Keyboard = decimal.Parse(property);
 
     property = config["PeripheralCosts:Mouse"];
     if (property is null)
         throw new ArgumentNullException("PeripheralCosts:Mouse not found in appsettings.json");
-    _costs.Mouse = float.Parse(property);
+    _costs.Mouse = decimal.Parse(property);
 
     property = config["PeripheralCosts:Chair"];
     if (property is null)
         throw new ArgumentNullException("PeripheralCosts:Chair not found in appsettings.json");
-    _costs.Chair = float.Parse(property);
+    _costs.Chair = decimal.Parse(property);
 }
 
 void GenerateDBScript(string _scriptOutput)
